requirements.txt
discord.py
python-dotenv
requests
schedule
flask
gunicorn